+++ application/src/app/gpus/gpu.component.html
@@ -1,3 +1,54 @@
-<div *ngIf="gpu">
-  {{ gpu.properties.deviceName }}
+<div class="main-container">
+  <main class="main">
+  <div>
+    <h1>This is a gpu.</h1>
+    <ul class="nav nav-tabs">
+      <li class="active">
+        <a data-toggle="tab" href="#device">
+          Device
+        </a>
+      </li>
+      <li>
+        <a data-toggle="tab" href="#features">
+          Features
+        </a>
+      </li>
+      <li>
+        <a data-toggle="tab" href="#limits">
+          Limits
+        </a>
+      </li>
+      <li>
+        <a data-toggle="tab" href="#extensions">
+          Extensions
+        </a>
+      </li>
+      <li>
+        <a data-toggle="tab" href="#formats">
+          Formats
+        </a>
+      </li>
+      <li>
+        <a data-toggle="tab" href="#queuefamilies">
+          Queue families
+        </a>
+      </li>
+      <li>
+        <a data-toggle="tab" href="#memory">
+          Memory
+        </a>
+      </li>
+      <li>
+        <a data-toggle="tab" href="#surface">
+          Surface
+        </a>
+      </li>
+      <li>
+        <a data-toggle="tab" href="#layers">
+          Layers
+        </a>
+      </li>
+    </ul>
+  </div>
+  </main>
 </div>
